Se realizo un cambio en la estructura haciendo que se heredara de una clase superior para utilizar metodos iguales en clases hijas
