+++ Proyecto Lenguajes Formales y Automatas/Proyecto Lenguajes Formales y Automatas.csproj
@@ -46,6 +46,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Automaton.cs" />
     <Compile Include="DFA.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
